+++ pyv/models/model.py
@@ -33,6 +33,17 @@ class Model:
         # Register module as global top module
         self.top = mod
 
+    def setProbes(self, probes: list[str]):
+        """Setup probes for ports.
+
+        Only ports whose full hierarchical name match at least one element from
+        `probes` will be logged during simulation.
+
+        Args:
+            probes (list[str]): List of strings to match ports to probe
+        """
+        self.sim.setProbes(probes)
+
     def run(self, num_cycles=1):
         """Runs the simulation.